Fix: make release-monitor statuses consistent (fixes #49)
This updates the `release-monitor` status check to leave the correct success status on semver-patch PRs when entering post-release mode. This allows the plugin logic to be simpler, because the appropriate action for any given PR only depends on whether the repo is in post-release mode, and it doesn't depend on what triggered the current event.