Fix: avoid reporting spurious for multiline commits referencing issues (#86)
Previously, if a PR had a single commit, and that commit's message spanned multiple lines, and the first line ended with (fixes #<number>), then the bot would report that the commit message didn't follow the guidelines because the issue reference wouldn't be at the "end" of the commit message. This updates the commit message checker to verify that issue references are at the end of the first line, rather than at the end of the entire commit message.